+++ patricare-back/src/modules/user/dto/user-2-document.dto.ts
@@ -1,4 +1,4 @@
-import { IsString, IsNumber } from 'class-validator';
+import { IsNumber } from 'class-validator';
 
 export class CreateUser2DocumentDto {
   @IsNumber() userId: number;
